+++ src/lib.rs
@@ -5,17 +5,14 @@
 #![warn(dead_code)]
 #![warn(missing_copy_implementations)]
 
-#[macro_use]
-extern crate log;
-
+mod channel;
 pub mod helper;
 mod middleware;
-mod requests;
 mod server;
 
 pub use actix_web::{
     error::PayloadError, http::header::HeaderMap, web::Payload, HttpMessage, HttpRequest,
     HttpResponse,
 };
-pub use requests::{Request, RequestReceiver};
+pub use http; // re-export http crate
 pub use server::{new, TestServer};