+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* Copy of Request is based on crate *http*
 * `TestServer::new` first argument is ToSocketAddrs
 * header function exports actix HeaderMap
 * added helper function to read response/request bodies